+++ web/plugins/eslint/rules/consistent-placeholders.js
