+++ web/plugins/eslint/rules/no-legacy-namespace-prefix.js
