+++ web/plugins/eslint/rules/require-ns-option.js
